:alembic: reaquainted myself with working with material-ui icons
