+++ react-dapp/package.json
@@ -4,7 +4,11 @@
   "private": true,
   "dependencies": {
     "@material-ui/core": "^3.9.3",
+    "@material-ui/icons": "^3.0.2",
+    "@material-ui/styles": "^3.0.0-alpha.10",
+    "@material/top-app-bar": "^1.1.1",
     "fortmatic": "^0.7.0",
+    "material-icons": "^0.3.1",
     "react": "^16.8.6",
     "react-content-loader": "^4.2.1",
     "react-dom": "^16.8.6",
