+++ react-dapp/src/components/Example/Example.js
@@ -1,6 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { withStyles } from "@material-ui/core/styles";
+import { styled } from "@material-ui/styles";
 import Button from "@material-ui/core/Button";
 
 const styles = theme => ({
@@ -12,10 +13,21 @@ const styles = theme => ({
   }
 });
 
+const MyButton = styled(Button)({
+  background: "linear-gradient(45deg, #FE6B8B 30%, #FF8E53 90%)",
+  border: 0,
+  borderRadius: 3,
+  boxShadow: "0 3px 5px 2px rgba(255, 105, 135, .3)",
+  color: "white",
+  height: 48,
+  padding: "0 30px"
+});
+
 function Example(props) {
   const { classes } = props;
   return (
     <div>
+      <MyButton>Custom Button</MyButton>
       <Button variant="contained" className={classes.button}>
         Default
       </Button>
