+++ react-dapp/src/components/Example/index.js
@@ -1,2 +1,1 @@
 export { default } from "./Example";
-export * from "./Example";
