+++ react-dapp/src/components/Header/index.js
@@ -1,19 +1,1 @@
-import React from "react";
-
-import Example from "../Example";
-
-import "./index.css";
-
-function Header() {
-  return (
-    <div>
-      <header className="App-header">
-        {/* <img src={logo} className="App-logo" alt="logo" /> */}
-        <p>Library of Babel</p>
-        <Example />
-      </header>
-    </div>
-  );
-}
-
-export default Header;
+export { default } from "./Header";
